Update remote models

+++ app/src/main/java/com/chebdowski/pokerdemo/data/model/Brings.kt
@@ -5,6 +5,6 @@ import com.fasterxml.jackson.annotation.JsonProperty
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 data class Brings(
-    @JsonProperty("min") val min: Int,
-    @JsonProperty("max") val max: Int
+    @JsonProperty("min") val min: Long,
+    @JsonProperty("max") val max: Long
 )
